requirements.txt
notebook
python-docx
PyPDF2
pdfplumber
pytesseract
pdf2image
Pillow
transformers
torch
sentencepiece
Flask
